exercise 9.4 and tests verified

+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/Food2DoorOrderingSystem/Product.java
@@ -7,7 +7,7 @@ public class Product {
     private String unit;
     private String productID;
 
-    public Product(String name, int quantity, String unit, String productID) {
+    public Product(String name, double quantity, String unit, String productID) {
         this.name = name;
         this.quantity = quantity;
         this.unit = unit;
